Several corrections and improvements to the detection of the various VC++ versions.
This also takes care of adding the Windows Kit (SDK) to the PATH in order to
enable rc.exe and other utilities to run. This appeared to affect VS2015 but
not newer VS versions. It may, however, affect older versions, too. But those
should as well be covered by these changes.